refactoring: routing, ssr, pages
