+++ rs-react-app/src/App.tsx
@@ -1,23 +1,10 @@
-import './styles/general/App.scss'
 
-import Header from './components/header/header';
-import Footer from './components/footer/footer';
-import Navigation from './components/navigation/navigation';
-import FlyoutElement from './components/flyoutElement/flyoutElement';
-
-import PageWrapper from './components/pageWrapper/pageWrapper';
 import ContentBlock from './components/contentBlock/contentBlock';
 
 const App = () => {
   return (
     <div data-testid="app-component">
-      <Header></Header>
-      <PageWrapper>
-        <Navigation></Navigation>
-        <FlyoutElement></FlyoutElement>
-        <ContentBlock></ContentBlock>
-      </PageWrapper>
-      <Footer></Footer>
+      <ContentBlock></ContentBlock>
     </div>
   );
 }
