+++ rs-react-app/src/app/layout.tsx
@@ -1,4 +1,13 @@
 import type { Metadata } from 'next'
+
+import '../../src/styles/general/App.scss'
+
+import Header from '../components/header/header'
+import Footer from '../components/footer/footer'
+import PageWrapper from '../components/pageWrapper/pageWrapper'
+import Navigation from '../components/navigation/navigation'
+
+import ProviderWrapper from '../components/flyoutWrapper/flyoutWrapper'
  
 export const metadata: Metadata = {
   title: 'My App',
@@ -13,7 +22,16 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body>
-        <div id="root">{children}</div>
+        <>
+          <Header></Header>
+          <PageWrapper>
+            <ProviderWrapper>
+              <Navigation></Navigation>
+              <div id="root">{children}</div>
+            </ProviderWrapper>
+          </PageWrapper>
+          <Footer></Footer>
+        </>
       </body>
     </html>
   )
