+++ rs-react-app/src/components/context/themeContext.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { createContext, useState, useEffect } from 'react';
 import type { ReactNode } from 'react';
