+++ rs-react-app/src/components/errorBoundary/errorBoundary.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { type ReactNode } from 'react';
 
 import PageWrapper from '../pageWrapper/pageWrapper';
