+++ rs-react-app/src/components/errorBoundaryMsg/errorBoundaryMsg.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React from "react";
 
 interface ErrorBoundaryMsgInterface {
