+++ rs-react-app/src/components/errorBtn/errorBtn.tsx
@@ -1,4 +1,4 @@
-
+"use client"
 interface ErrorBtnProps {
   onClick: () => void
 }
