+++ rs-react-app/src/components/flyoutElement/flyoutElement.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import type { AppDispatch, RootState } from '../../store';
