+++ rs-react-app/src/components/footer/footer.tsx
@@ -1,4 +1,3 @@
-
 const Footer = () => {
   return (
     <div className='footer' data-testid="footer">
