+++ rs-react-app/src/components/handleForm/handleForm.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useState } from 'react';
 
 interface handleFormProps {
