+++ rs-react-app/src/components/header/header.tsx
@@ -1,4 +1,3 @@
-
 const Header = () => {
   return (
     <>
