+++ rs-react-app/src/components/navigation/navigation.tsx
@@ -1,5 +1,4 @@
-
-import { Link } from 'react-router';
+import Link from 'next/link';
 
 import ThemeChange from '../themeChange/themeChange';
 
@@ -7,9 +6,9 @@ const Navigation = () => {
   return (
     <div data-testid='navigation-test' className='navigation'>
       <h2>Navigation</h2>
-      <Link to="/">Go to main</Link>
-      <Link to="/about">Go to about</Link>
-      <Link to="/404">Go to 404</Link>
+      <Link href="/">Go to main</Link>
+      <Link href="/about">Go to about</Link>
+      <Link href="/404">Go to 404</Link>
       <ThemeChange></ThemeChange>
     </div>
   );
