+++ rs-react-app/src/components/themeChange/themeChange.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { useContext } from 'react';
 
 import { ThemeContext } from '../context/themeContext';
