+++ rs-react-app/src/components/usePaginationHook/usePaginationHook.tsx
@@ -1,20 +1,22 @@
+"use client"
 import { useEffect } from 'react';
 
 export function usePaginationHook(
   param: number | string | undefined,
   searchParams: URLSearchParams,
-  updateSearchParams: (params: URLSearchParams) => void
+  routerPush: (url: string) => void
 ) {
 
   useEffect(() => {
     if (param === null || param === '') return;
+
+    const newParams = new URLSearchParams(searchParams.toString())
 
     if(typeof(param) === 'number') {
       const newParams = new URLSearchParams(searchParams)
       newParams.delete('q')
       newParams.set('page', `${String(param)}`) 
       newParams.set('details', '1')
-      updateSearchParams(newParams)
     } 
     else {
 
@@ -22,8 +24,9 @@ export function usePaginationHook(
       newParams.delete('page')
       newParams.delete('details')
       newParams.set('q', String(param))
-      updateSearchParams(newParams)
     }
 
-  }, [param, searchParams, updateSearchParams]);
+    routerPush(`${newParams.toString()}`)
+
+  }, [param, searchParams, routerPush]);
 }
