+++ rs-react-app/src/main.tsx
@@ -1,28 +1,28 @@
-import { createRoot } from 'react-dom/client'
-import { BrowserRouter, Routes, Route } from 'react-router'
+// import { createRoot } from 'react-dom/client'
+// import { BrowserRouter, Routes, Route } from 'react-router'
 
-import App from './App'
-import About from './components/about/about'
-import ErrorPage from './components/404/404'
+// import App from './App'
+// import About from './components/about/about'
+// import ErrorPage from './components/404/404'
 
-import ErrorBoundary from './components/errorBoundary/errorBoundary'
+// import ErrorBoundary from './components/errorBoundary/errorBoundary'
 
-import { ThemeProvider } from './components/context/themeContext'
-import { Provider } from 'react-redux'
-import { store } from './store'
+// import { ThemeProvider } from './components/context/themeContext'
+// import { Provider } from 'react-redux'
+// import { store } from './store'
 
-createRoot(document.getElementById('root')!).render(
-  <BrowserRouter>
-    <ErrorBoundary >
-      <Provider store={store}>
-        <ThemeProvider>
-          <Routes>
-            <Route path="/" element={<App />} />
-            <Route path="/about" element={<About />} />
-            <Route path="*" element={<ErrorPage />} />
-          </Routes>
-        </ThemeProvider>
-      </Provider>
-    </ErrorBoundary>
-  </BrowserRouter>
-)
+// createRoot(document.getElementById('root')!).render(
+//   <BrowserRouter>
+//     <ErrorBoundary >
+//       <Provider store={store}>
+//         <ThemeProvider>
+//           <Routes>
+//             <Route path="/" element={<App />} />
+//             <Route path="/about" element={<About />} />
+//             <Route path="*" element={<ErrorPage />} />
+//           </Routes>
+//         </ThemeProvider>
+//       </Provider>
+//     </ErrorBoundary>
+//   </BrowserRouter>
+// )
